requirements.txt
supabase 
python-dotenv 
lxml 
python-dateutil 
pytz 
loguru
streamlit
pandas
plotly
